Create integration test directory
The remainder of the test modules that aren't unit tests have been
moved into this new `integration` directory. I figured that it would be
accurate to describe these as integration tests since they test the
token tracker as a whole, and focus more on specific common use-cases
rather than on covering the API of the tracker itself.

The main motivation behind this is that I want to create a module for
test utility functions, and it would be nice if this was kept separate
from the tests themselves.