Create token unit tests (#29)
The `createSimpleTokenEnvironment` function was extracted from the
`simple-token` tests into a helper library, and reused for the token
unit tests.

The token unit tests focused on the constructor specifically, using the
`serialize` function to assert expectations about the token state. The
rest of the module has not yet been tested, and neither have conditions
that are known to make the token fail upon construction. The failure
conditions will be tested in a later PR; at the moment it's difficult
to test them because they throw an exception asynchronously, after the
constructor returns.