Move initial token update into tracker
Previously the Token constructor had triggered an asynchronous update
of the token metadata and balance. Unfortunately this update wouldn't
trigger either an `update` after completion or an `error` event after
failure, leaving the user of the tracker unable to handle the error
and unable to tell when the metadata update had finished.

This update has been moved into the Token Tracker instead, and now
emits both `update` and `error` events as appropriate.

The Token `serialize` method return value has changed to return the
serialized token rather than an array of symbol, balance, and decimal.
This was done because it was slightly more convenient, and the old
return value was never used anywhere anyway.