Ensure first token update is always emitted
The first token update was sometimes not emitted because the initial
token state happened to be the same as the state after the update. This
was particularly problematic with the balance, as the tracker will
initialize each token with a balance of zero even if none was
specified.

The initial token state has been removed, which ensures the first
update will always be emitted.

I considered skipping the initial update event if all of the token data
was fully specified from the beginning (including the balance), but
decided it wasn't worthwhile. The token data includes a derived field
anyway (the `string` field) which can't be specified, so technically
it has always changed at least with that field. If we were interested
in doing this in the future but preserving the token defaults (balance
and decimals set to zero), we'd have to add a separate serialization
method to the token module that doesn't apply those defaults.

Fixes #14