Prevent redundant update caused by simultaneous token updates (#32)
The tracker attempts to compare the previous token balances to the
current token balance when deciding whether to emit an `update` event.
However, the previous balance it compares against is the balance before
the update, not last balance emitted. If a second token update is
started before the first one finishes, both will be comparing using the
same previous value, so both might result in the same `update` event
being emitted.

This redundant `update` event can occur when some token balance has
changed, and two token updates are triggered in quick succession. This
could happen if two blocks were mined in quick succession, or if two
updates were manually triggered, or if a block was mined and an update
triggered at around the same time.

This has been prevented by caching the token state emitted by the most
recent `update` event. This cached state is used to determine whether
the token state has changed, rather than the state before the update
began.

The state is initialized prior to the first update to remain consistent
with the previous behavior. This will be removed in a subsequent PR.