Add test to ensure initial update is emitted
It was difficult to test that the initial update was emitted because
an update was also triggered immediately after construction, as a side-
effect of adding a listener to the block tracker.

It has now been tested by making a transfer immediately after the
tracker is constructed. By the time the tracker is initiated, the
initial token update had already been fired, so this results in two
separate updates. Previously this would have been a single update.