Emit events for initial token updates (#34)
Previously the Token constructor had triggered an asynchronous update
of the token metadata and balance. Unfortunately this update wouldn't
trigger either an `update` after completion or an `error` event after
failure, leaving the user of the tracker unable to handle the error
and unable to tell when the metadata update had finished.

This update has been moved into the Token Tracker instead, and now
emits both `update` and `error` events as appropriate.

The change to the `update` event has been tested by making a transfer
immediately after the tracker is constructed. By the time the tracker is
initiated, the initial token update had already been fired, so this results in
two separate updates. Previously this would have been a single update.

The Token `serialize` method return value has changed to return the
serialized token rather than an array of symbol, balance, and decimal.
This was done because it was slightly more convenient, and the old
return value was never used anywhere anyway.