Ignore extraneous files when publishing (#36)
The files to be included in the final bundle are now specified using
the `files` property in the manifest. This field includes a set of
default files (such as the LICENSE and manifest) along with the
specified files, and excludes all other files.

In this case, it means the tests, examples, and the CI config is no
longer published.

For now the original source is published in addition to the compiled
source, so that either may be used by consuming packages. We should
probably include both until we decide to remove the build step.