Add allow-scripts (#118)
Adding `@lavamoat/allow-scripts` and `@lavamoat/preinstall-always-fail`
is a prerequisite to upgrading to Yarn v3.

If you want to run tests or use this library locally, you will now need
to run `yarn setup` instead of `yarn install`. This is temporary and
the setup command will be changed back to `yarn install` when Yarn is
upgraded to v3.